Create SPATIAL_FIT_2DFUNC.f
This is a code that can be used along with the Subaru/COMICS data reduction pipeline. It basically just does a 2D fit, but it's larger role is to help determine the spatially constant axis in the spectroscopic data. (The target traces out a curve during the spectral observation,so we have to find a function that fits this curve, so that we can properly extract the data. For reference we use the curve traced out by the guide star, and then use the fitting results for the scientific observation- in the current case, this means IRAS18434-0242)