Rewriting to use ready-made COMQ files
Sakon-san pointed out that the data appeared to be made from only a single frame of the COMA files.

Also that the way I had been making the COMQ files was wrong. However it seems that the already-made COMQ files are available from the Subaru database, so he kindly provided those. We'll proceed with the pre-prepared COMQ files, and remove the step where I was (incorrectly) making COMQsA