Another correction to q_startrace parameters
Since updating q_startrace to use the COMQ images, it's clear that the "X-range" (3rd parameter) isn't intended to be a guess. The data should span the entire X-range, since that's the direction of dispersion. The Y-range parameter however, is a guess of where the peak should be.

So we'll change the X part back to '1-320'. As for the Y-part, well, that needs to be adjusted too- but it'll depend on the image. updating that later....