Update q_all.sh
Added several comments after reviewing the code with Sakon-san, for changes which needed to be made soon

Basically, there is some mismatch between the way the frames in the fits files are ordered in THIS observation vs. PAST observations.
Thus we were essentally trying to make FLAT FIELD corrections based on READ-OUT noise and vice versa, resulting in very strange artifacts in the images.

It also appears that the Q18.8 micron data is in fact missing from this observation set.

Also, some of the dark images may have mistaken timings, so we have to be sure to only use the "88297" and "88295" images.

Anyways, the major point of the day is this: for the q_list_stat commands -> "2" refers to the flat data, and "1" refers to the readout noise.