COMA vs. COMQ for Flat processing, correction
Hopefully a final clarification: it seems that with the DOMEFLAT images - as used for the NL spectrscopy, rather than the SKYFLATS -  we use the COMA images! 
 
Which should make sense, since the COMQ processing tries to isolate the object. In flat-fielding, it doesn't really make sense to isolate the "object". Which raises a question for later- why do we even have COMQs of DOMEFLATS?