Revised comments;　Removed redundant commands.
I found that the same dark file is used throughout the "q_list_stat" portion of th SKY image making. Thus I moved the dark-file-scaling command to the top of that process, so that the same "FDARK1" temporary file can be used throughout the process, rather than re-calculating it for each image.