Adding the beam separation step to the SKY image construction process
"q_bsep" sepearates the beam-images in the COMA files... These "beams" correspond to the chopping motion during the observation- chopping is the rapid "on-off" switching done during the exposures. The point is to get data both directly ON the target, and just OFF the target. The OFF beam can then be used to make a SKY image- and that's what we have to do now. If we dont separate the beams, then our SKY image will be all confused with the target image.