Added FLAT correction to the SKY image making steps..
I am not yet entirely certain if this is where the FLAT correction should go, but according to the manual, it appears that way. It makes sense that we would need to both subtract dark current and carry-out flat fielding even when making the SKY images (since NL uses the DOME flat).

I thought to try this upon noticing that the same pattern appearing in the FLAT images seemed to appear in the SKY images. And the previously generated SKY images couldn't pass through the q_sky_nlow process, so I assumed something was wrong with them. After flat correction, the pattern is gone and the process completes (q_sky_nlow iterates through all the Y points requested).

I still need to confirm that the changes made in this commit are the appropriate method...