Update q_all.sh
Now making another sub-version which attempts to run Q_SKY_NLOW without flat correction AND without Q_BSEP. Well, essentially, this equates to running the code just as it was in the original attempts. The major difference now is actually in Q_SKY_NLOW itself, and that I've added some diagnostic echoes. If this works, then we can more this version with the master branch and close these sub-versions!