Update q_all_2.sh
Added commands for the final processing into "spectra files", or ".SPC" format. 
WARNING: These latest commands are essentially copied and pasted from Sakon-san's memo, since we are now working with the same number of images as the memo. So be careful to make sure all of the parameters are the appropriate ones!!!!