Update q_all_2.sh
Since the failed NL_OBJ_B07N.fits GEOTRAN process and the error messages for the other B**N images seem to come from a bad fit of the constant spatial axis (now fixed), I removed the comment about this problem. 

But the reminder to check the output stays! For the future, I think it will be nice to automatically display the output using something that's a bit lighter than DS9 - just for the sake of having a quick look. DS9 is very slow to load when connecting remotely.

Perhaps if we can move this process to a notebook environment (i.e. ipynb), it would be easier to organize the processing steps and output checks. However perhaps we'll need to wrap the q_*** codes, which are written in C. They're very fast, but perhaps they wouldn't be that way if we switched to python routines. 

I think just moving the "top-level" of the processing to Python, and using astropy visualization tools would be nice.