Create q_interp.py
This is a code for interpolating new values for some of the bad pixels in the NL IRAS18434-0242 data.

Most likely we'll need several processes to deal with all of the bad pixels, since there are a few types of them:

Lone "hot" pixels - a single pixel which has an extremely high value compared to its neighbors

Lone "dead" pixels- the same as lone hot pixels, but with extremely low or negative values

Patches: Either "hot" or "dead", these are clusters of bad pixels. Usually they occur in blocks of four pixels. Sometimes their positions can vary across exposures, so when you run q_combine on a bunch of exposures (like, 16, in the current case), you might notice that the patches "grow".

I think that the first two cases can be done by a simple nearest-neighbor interpolation, perhaps like this one in scipy:
   http://docs.scipy.org/doc/scipy-0.17.0/reference/generated/scipy.interpolate.RegularGridInterpolator.html