Update q_interp.py
Based on some closer inspection of the images and reading up on numpy/scipi etc. interpolation suggestions, I decided to break this problem down into a 1D process:

There's a lot of variation from pixel to pixel in the Y-dimension, but far less pixel-to-pixel variation along the X-dimension. The glaring exception is  X-positions that correspond to line emission- however there does not appear to be any bad pixels that fall on a line emission profile.

With that in mind, we'll:
1) read each row one by one
2) identify the "bad" positions
3) interpolate "new" values for the bad positions, using np.interp

1) and 3) are simple enough: the tricky part is identifying which pixels are bad. They are not labled as NaN, and there is no mask.

There a few patterns that they appear to follow, however: all pixels with a value less than -130. have the appearance of a "dead pixel". Pixels with values this low are surrounded by positive-valued pixels, except in the cases where the dead pixels occur in 2x2p blocks. The blocks themselves follow the same pattern though- the 12 pixels on the border are positive-valued, while the block pixels are very negative.

This means that we can at least proceed with the first wave of fixes: label the stranded, very negative pixels as NaN- then replace these with interpolated values.

The next step will be to determine a proper threshold for the "hot" pixels, which should be a similar process but may not a bit more thought. For now I'm implementing this in a serious of "passes". First I deal with the brightest pixels in the image, since these are brighter than even the brightest source pixels.

In the next two passes, I deal with the brightest isolated pixels/pixel-patches which are outside of the "source data" regions.

In the third pass, since the hot pixels are close to the level of the brightest sky emission lines, I also exclude the bright sky-dominated regions.